+++ serialization/dotnet/json/Microsoft.Kiota.Serialization.Json.Tests/Mocks/TestEntity.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Microsoft.Kiota.Abstractions;
 using Microsoft.Kiota.Abstractions.Serialization;
 
 namespace Microsoft.Kiota.Serialization.Json.Tests.Mocks
@@ -13,7 +14,17 @@ namespace Microsoft.Kiota.Serialization.Json.Tests.Mocks
         /// <summary>Read-only.</summary>
         public TestEnum? Numbers { get; set; }
         /// <summary>Read-only.</summary>
+        public TimeSpan? WorkDuration { get; set; }
+        /// <summary>Read-only.</summary>
+        public Date? BirthDay { get; set; }
+        /// <summary>Read-only.</summary>
+        public Time? StartWorkTime { get; set; }
+        /// <summary>Read-only.</summary>
+        public Time? EndWorkTime { get; set; }
+        /// <summary>Read-only.</summary>
         public DateTimeOffset? CreatedDateTime { get; set; }
+        /// <summary>Read-only.</summary>
+        public string OfficeLocation { get; set; }
         /// <summary>
         /// Instantiates a new entity and sets the default values.
         /// </summary>
@@ -30,6 +41,11 @@ namespace Microsoft.Kiota.Serialization.Json.Tests.Mocks
                 {"id", (o,n) => { (o as TestEntity).Id = n.GetStringValue(); } },
                 {"numbers", (o,n) => { (o as TestEntity).Numbers = n.GetEnumValue<TestEnum>(); } },
                 {"createdDateTime", (o,n) => { (o as TestEntity).CreatedDateTime = n.GetDateTimeOffsetValue(); } },
+                {"officeLocation", (o,n) => { (o as TestEntity).OfficeLocation = n.GetStringValue(); } },
+                {"workDuration", (o,n) => { (o as TestEntity).WorkDuration = n.GetTimeSpanValue(); } },
+                {"birthDay", (o,n) => { (o as TestEntity).BirthDay = n.GetDateValue(); } },
+                {"startWorkTime", (o,n) => { (o as TestEntity).StartWorkTime = n.GetTimeValue(); } },
+                {"endWorkTime", (o,n) => { (o as TestEntity).EndWorkTime = n.GetTimeValue(); } },
             };
         }
         /// <summary>
@@ -42,6 +58,11 @@ namespace Microsoft.Kiota.Serialization.Json.Tests.Mocks
             writer.WriteStringValue("id", Id);
             writer.WriteEnumValue<TestEnum>("numbers",Numbers);
             writer.WriteDateTimeOffsetValue("createdDateTime", CreatedDateTime);
+            writer.WriteStringValue("officeLocation", OfficeLocation);
+            writer.WriteTimeSpanValue("workDuration", WorkDuration);
+            writer.WriteDateValue("birthDay", BirthDay);
+            writer.WriteTimeValue("startWorkTime", StartWorkTime);
+            writer.WriteTimeValue("endWorkTime", EndWorkTime);
             writer.WriteAdditionalData(AdditionalData);
         }
     }