+++ CHANGELOG.md
@@ -9,9 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added the ability to configure the underlying transport in Go. #1003
+- Added additional date time (date, time, duration) types in the generation process. #1017
+- PHP Request Adapter (includes middleware) #1048, #918, #1024, #1025
+- Added support for PHP Json Serialization.
+- Update PHP abstractions to make property keys and values nullable in `SerializationWriter.php`.
+
 ### Changed
 
 - Fixed an issue where multiple api clients could run into racing conditions in Go.
+- Fixed a bug where empty additional data in Go would lead to invalid JSON payloads during serialization.
+- Fixed a bug where Go serialization would write empty arrays for nil values.
+- Modified the TypeScript RequestInformation URL paramater data type from URL to string.
+- Modified TypeScript packages to provide CJS and ESM modules.
+- Modified the TypeScript RequestInformation query and path paramaters data type from Map to Record Type.
+- Modified TypeScript RequestInformation headers and options to Record type.
+- Modified the TypeScript RequestInformation content data type to ArrayBuffer.
 
 ## [0.0.15] - 2021-12-17
 
@@ -33,6 +46,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixes a bug in Go where empty collections would not be serialized.
 - Fixes a bug where generation would fail because of empty usings.
 - Fixes a bug where Java and Go escaped model properties would not serialize properly.
+- Fixes a bug where null values would not be added to additionalData if there was no matching property in dotnet. 
+- Fixes a bug where deserialzation of enums would throw an ArgumentExcpetion if the member didn't exist in dotnet.
 
 ## [0.0.14] - 2021-11-08